pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "filebot"
version = "0.1.0"
description = "A Simple langchain app to organize my pdfs"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.1.14",
    "langchain-community>=0.0.32",
    "pypdf>=4.0.1",
    "python-dotenv>=1.0.0",
    "ollama>=0.1.25",
    "requests>=2.31.0",
    "streamlit>=1.28",
    "langchain-core>=1.2.9",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.4",
    "pytest-cov==4.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]
